+++ src/widgets/simulator.rs
@@ -131,7 +131,7 @@ impl Simulator<'_> {
                         )
                         .text(progress_bar_text(
                             state.durability,
-                            i16::from(self.settings.max_durability),
+                            self.settings.max_durability,
                         ))
                         .corner_radius(0),
                     );
